Move OSS license-handling to non-ENT file
